Add files via upload
This was a worksheet I created via ChatGPT to drill in my learning from the DataCamp course I am currently doing. Some of these I did get wrong in which I made note of for my next worksheet to solidify my understanding of lists in python.

I struggled mainly with syntax and have made notes on how to correct this going forwards.